+++ include/circt/Dialect/MSFT/CMakeLists.txt
@@ -14,7 +14,6 @@ mlir_tablegen(MSFTEnums.h.inc -gen-enum-decls)
 mlir_tablegen(MSFTEnums.cpp.inc -gen-enum-defs)
 mlir_tablegen(MSFTAttributes.h.inc -gen-attrdef-decls)
 mlir_tablegen(MSFTAttributes.cpp.inc -gen-attrdef-defs)
-mlir_tablegen(MSFTPasses.h.inc -gen-pass-decls)
 
 set(LLVM_TARGET_DEFINITIONS MSFTOpInterfaces.td)
 mlir_tablegen(MSFTOpInterfaces.h.inc -gen-op-interface-decls)
@@ -22,6 +21,9 @@ mlir_tablegen(MSFTOpInterfaces.cpp.inc -gen-op-interface-defs)
 
 add_public_tablegen_target(MLIRMSFTExtraIncGen)
 
-add_circt_doc(MSFT MSFTPasses -gen-pass-doc)
+set(LLVM_TARGET_DEFINITIONS MSFTPasses.td)
+mlir_tablegen(MSFTPasses.h.inc -gen-pass-decls)
+add_public_tablegen_target(CIRCTMSFTTransformsIncGen)
+add_circt_doc(MSFTPasses MSFTPasses -gen-pass-doc)
 
 add_dependencies(circt-headers MLIRMSFTExtraIncGen)
