+++ include/circt/Dialect/MSFT/MSFTPasses.h
@@ -9,14 +9,25 @@
 #ifndef CIRCT_DIALECT_MSFT_MSFTPASSES_H
 #define CIRCT_DIALECT_MSFT_MSFTPASSES_H
 
+#include "circt/Dialect/HW/HWOps.h"
 #include "circt/Dialect/MSFT/MSFTOps.h"
 
-#include "circt/Dialect/HW/HWOpInterfaces.h"
-#include "circt/Dialect/HW/HWOps.h"
+#include "mlir/Pass/PassRegistry.h"
+
+namespace mlir {
+class Pass;
+} // namespace mlir
 
 namespace circt {
 namespace msft {
-void registerMSFTPasses();
+
+std::unique_ptr<mlir::Pass> createWireCleanupPass();
+std::unique_ptr<mlir::Pass> createPartitionPass();
+std::unique_ptr<mlir::Pass> createLowerToHWPass();
+std::unique_ptr<mlir::Pass> createLowerInstancesPass();
+std::unique_ptr<mlir::Pass> createLowerConstructsPass();
+std::unique_ptr<mlir::Pass> createExportTclPass();
+std::unique_ptr<mlir::Pass> createDiscoverAppIDsPass();
 
 /// A set of methods which are broadly useful in a number of dialects.
 struct PassCommon {
@@ -39,6 +50,11 @@ protected:
                                 SmallVectorImpl<hw::HWModuleLike> &mods,
                                 DenseSet<Operation *> &modsSeen);
 };
+
+/// Generate the code for registering passes.
+#define GEN_PASS_REGISTRATION
+#include "circt/Dialect/MSFT/MSFTPasses.h.inc"
+
 } // namespace msft
 } // namespace circt
