+++ lib/Dialect/MSFT/Transforms/MSFTToHW.cpp
@@ -1,0 +1,231 @@
+//===- MSFTToHW.cpp - MSFT lowering pass ------------------------*- C++ -*-===//
+//
+// Part of the LLVM Project, under the Apache License v2.0 with LLVM Exceptions.
+// See https://llvm.org/LICENSE.txt for license information.
+// SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
+//
+//===----------------------------------------------------------------------===//
+
+#include "PassDetails.h"
+#include "circt/Dialect/HW/HWAttributes.h"
+#include "circt/Dialect/HW/HWOps.h"
+#include "circt/Dialect/HW/HWTypes.h"
+#include "circt/Dialect/MSFT/ExportTcl.h"
+#include "circt/Dialect/MSFT/MSFTDialect.h"
+#include "circt/Dialect/MSFT/MSFTOpInterfaces.h"
+#include "circt/Dialect/MSFT/MSFTOps.h"
+#include "circt/Dialect/MSFT/MSFTPasses.h"
+#include "circt/Dialect/SV/SVOps.h"
+#include "circt/Dialect/Seq/SeqOps.h"
+#include "circt/Support/Namespace.h"
+
+#include "mlir/IR/BlockAndValueMapping.h"
+#include "mlir/IR/BuiltinOps.h"
+#include "mlir/IR/PatternMatch.h"
+#include "mlir/Pass/Pass.h"
+#include "mlir/Pass/PassRegistry.h"
+#include "mlir/Transforms/DialectConversion.h"
+#include "mlir/Transforms/GreedyPatternRewriteDriver.h"
+
+using namespace circt;
+using namespace msft;
+
+//===----------------------------------------------------------------------===//
+// Lower MSFT to HW.
+//===----------------------------------------------------------------------===//
+
+namespace {
+/// Lower MSFT's InstanceOp to HW's. Currently trivial since `msft.instance` is
+/// currently a subset of `hw.instance`.
+struct InstanceOpLowering : public OpConversionPattern<InstanceOp> {
+public:
+  using OpConversionPattern::OpConversionPattern;
+
+  LogicalResult
+  matchAndRewrite(InstanceOp, OpAdaptor adaptor,
+                  ConversionPatternRewriter &rewriter) const final;
+};
+} // anonymous namespace
+
+LogicalResult
+InstanceOpLowering::matchAndRewrite(InstanceOp msftInst, OpAdaptor adaptor,
+                                    ConversionPatternRewriter &rewriter) const {
+  Operation *referencedModule = msftInst.getReferencedModule();
+  if (!referencedModule)
+    return rewriter.notifyMatchFailure(msftInst,
+                                       "Could not find referenced module");
+  if (!hw::isAnyModule(referencedModule))
+    return rewriter.notifyMatchFailure(
+        msftInst, "Referenced module was not an HW module");
+
+  ArrayAttr paramValues;
+  if (isa<hw::HWModuleExternOp>(referencedModule)) {
+    paramValues = msftInst.parametersAttr();
+    if (!paramValues)
+      paramValues = rewriter.getArrayAttr({});
+  } else {
+    auto instAppendParam = hw::ParamExprAttr::get(
+        hw::PEO::StrConcat,
+        {hw::ParamDeclRefAttr::get(rewriter.getStringAttr("__INST_HIER"),
+                                   rewriter.getNoneType()),
+         rewriter.getStringAttr("."), msftInst.sym_nameAttr()});
+    paramValues = rewriter.getArrayAttr(
+        {hw::ParamDeclAttr::get("__INST_HIER", instAppendParam)});
+  }
+
+  auto hwInst = rewriter.create<hw::InstanceOp>(
+      msftInst.getLoc(), referencedModule, msftInst.instanceNameAttr(),
+      SmallVector<Value>(adaptor.getOperands().begin(),
+                         adaptor.getOperands().end()),
+      paramValues, msftInst.sym_nameAttr());
+  hwInst->setDialectAttrs(msftInst->getDialectAttrs());
+  rewriter.replaceOp(msftInst, hwInst.getResults());
+  return success();
+}
+
+namespace {
+/// Lower MSFT's ModuleOp to HW's.
+struct ModuleOpLowering : public OpConversionPattern<MSFTModuleOp> {
+public:
+  ModuleOpLowering(MLIRContext *context, StringRef outputFile)
+      : OpConversionPattern::OpConversionPattern(context),
+        outputFile(outputFile) {}
+
+  LogicalResult
+  matchAndRewrite(MSFTModuleOp mod, OpAdaptor adaptor,
+                  ConversionPatternRewriter &rewriter) const final;
+
+private:
+  StringRef outputFile;
+};
+} // anonymous namespace
+
+LogicalResult
+ModuleOpLowering::matchAndRewrite(MSFTModuleOp mod, OpAdaptor adaptor,
+                                  ConversionPatternRewriter &rewriter) const {
+  if (mod.body().empty()) {
+    std::string comment;
+    llvm::raw_string_ostream(comment)
+        << "// Module not generated: \"" << mod.getName() << "\" params "
+        << mod.parameters();
+    // TODO: replace this with proper comment op when it's created.
+    rewriter.replaceOpWithNewOp<sv::VerbatimOp>(mod, comment);
+    return success();
+  }
+
+  ArrayAttr params = rewriter.getArrayAttr({hw::ParamDeclAttr::get(
+      rewriter.getStringAttr("__INST_HIER"),
+      rewriter.getStringAttr("INSTANTIATE_WITH_INSTANCE_PATH"))});
+  auto hwmod = rewriter.replaceOpWithNewOp<hw::HWModuleOp>(
+      mod, mod.getNameAttr(), mod.getPorts(), params);
+  rewriter.eraseBlock(hwmod.getBodyBlock());
+  rewriter.inlineRegionBefore(mod.getBody(), hwmod.getBody(),
+                              hwmod.getBody().end());
+
+  auto opOutputFile = mod.fileName();
+  if (opOutputFile) {
+    auto outputFileAttr = hw::OutputFileAttr::getFromFilename(
+        rewriter.getContext(), *opOutputFile, false, true);
+    hwmod->setAttr("output_file", outputFileAttr);
+  } else if (!outputFile.empty()) {
+    auto outputFileAttr = hw::OutputFileAttr::getFromFilename(
+        rewriter.getContext(), outputFile, false, true);
+    hwmod->setAttr("output_file", outputFileAttr);
+  }
+
+  return success();
+}
+namespace {
+
+/// Lower MSFT's ModuleExternOp to HW's.
+struct ModuleExternOpLowering : public OpConversionPattern<MSFTModuleExternOp> {
+public:
+  ModuleExternOpLowering(MLIRContext *context, StringRef outputFile)
+      : OpConversionPattern::OpConversionPattern(context),
+        outputFile(outputFile) {}
+
+  LogicalResult
+  matchAndRewrite(MSFTModuleExternOp mod, OpAdaptor adaptor,
+                  ConversionPatternRewriter &rewriter) const final;
+
+private:
+  StringRef outputFile;
+};
+} // anonymous namespace
+
+LogicalResult ModuleExternOpLowering::matchAndRewrite(
+    MSFTModuleExternOp mod, OpAdaptor adaptor,
+    ConversionPatternRewriter &rewriter) const {
+  auto hwMod = rewriter.replaceOpWithNewOp<hw::HWModuleExternOp>(
+      mod, mod.getNameAttr(), mod.getPorts(), mod.verilogName().value_or(""),
+      mod.parameters());
+
+  if (!outputFile.empty()) {
+    auto outputFileAttr = hw::OutputFileAttr::getFromFilename(
+        rewriter.getContext(), outputFile, false, true);
+    hwMod->setAttr("output_file", outputFileAttr);
+  }
+
+  return success();
+}
+
+namespace {
+/// Lower MSFT's OutputOp to HW's.
+struct OutputOpLowering : public OpConversionPattern<OutputOp> {
+public:
+  using OpConversionPattern::OpConversionPattern;
+
+  LogicalResult
+  matchAndRewrite(OutputOp out, OpAdaptor adaptor,
+                  ConversionPatternRewriter &rewriter) const final {
+    rewriter.replaceOpWithNewOp<hw::OutputOp>(out, out.getOperands());
+    return success();
+  }
+};
+} // anonymous namespace
+
+namespace {
+struct LowerToHWPass : public LowerToHWBase<LowerToHWPass> {
+  void runOnOperation() override;
+};
+} // anonymous namespace
+
+void LowerToHWPass::runOnOperation() {
+  auto top = getOperation();
+  auto *ctxt = &getContext();
+
+  // The `hw::InstanceOp` (which `msft::InstanceOp` lowers to) convenience
+  // builder gets its argNames and resultNames from the `hw::HWModuleOp`. So we
+  // have to lower `msft::MSFTModuleOp` before we lower `msft::InstanceOp`.
+
+  // Convert everything except instance ops first.
+
+  ConversionTarget target(*ctxt);
+  target.addIllegalOp<MSFTModuleOp, MSFTModuleExternOp, OutputOp>();
+  target.addLegalDialect<hw::HWDialect>();
+  target.addLegalDialect<sv::SVDialect>();
+
+  RewritePatternSet patterns(ctxt);
+  patterns.insert<ModuleOpLowering>(ctxt, verilogFile);
+  patterns.insert<ModuleExternOpLowering>(ctxt, verilogFile);
+  patterns.insert<OutputOpLowering>(ctxt);
+  patterns.insert<RemoveOpLowering<EntityExternOp>>(ctxt);
+  patterns.insert<RemoveOpLowering<DesignPartitionOp>>(ctxt);
+
+  if (failed(applyPartialConversion(top, target, std::move(patterns))))
+    signalPassFailure();
+
+  // Then, convert the InstanceOps
+  target.addDynamicallyLegalDialect<MSFTDialect>([](Operation *op) {
+    return isa<DynInstDataOpInterface, DeclPhysicalRegionOp,
+               InstanceHierarchyOp>(op);
+  });
+  RewritePatternSet instancePatterns(ctxt);
+  instancePatterns.insert<InstanceOpLowering>(ctxt);
+  if (failed(applyPartialConversion(top, target, std::move(instancePatterns))))
+    signalPassFailure();
+}
+
+std::unique_ptr<Pass> circt::msft::createLowerToHWPass() {
+  return std::make_unique<LowerToHWPass>();
+}
